Create ping.c
This is a C program for the ping. There are no system calls available for performing ping in Linux. So let's use this program instead. 

You will have to use root privilege to execute the compiled program as it uses raw socket. I don't think it will be an issue.

One of the assignment for the class is to modify this program for following:

1. Hostname should be taken from command line as an argument instead of fixed value (like 127.0.0.1)
2. Measure the response time in micro-seconds, i.e time between request was sent and reply was received.
3. Response timeout value can be passed as command line argument. Default will be 2 seconds.
4. Number of ping request can be passed as command line argument.  Default will be 5. 
5. Program should print following output:
"Host www.google.com is alive. Average response time is 125000 micro-seconds."
 If ping is not successful:
"Host www.google.com is not reachable".